+++ src/main/java/com/abelium/inatrace/components/product/ProductTypeService.java
@@ -33,7 +33,7 @@ public class ProductTypeService extends BaseService {
         return new ApiPaginatedResponse<>(new ApiPaginatedList<>(productTypeList, productTypeList.size()));
     }
 
-    public ApiResponse<ApiProductType> getProductType(Integer id) throws ApiException {
+    public ApiResponse<ApiProductType> getProductType(Long id) throws ApiException {
         ProductType productType = em.find(ProductType.class, id);
 
         if (productType == null) {
@@ -62,7 +62,7 @@ public class ProductTypeService extends BaseService {
     }
 
     @Transactional
-    public ApiDefaultResponse deleteProductType(Integer id) throws ApiException {
+    public ApiDefaultResponse deleteProductType(Long id) throws ApiException {
         if (id == null) {
             throw new ApiException(ApiStatus.INVALID_REQUEST, "Id must be provided!");
         }
